/!\ Important Update²
Just revising a bit of logic, and updating comments, because I forgot to do it in the previous commit of the update.
It's admittedly better to work with ranges than with "starting from there, do that". I can't believe that looking at the code the first time, I splurged out comments working with "above said temp", which was accurate with what the code was saying... though said code didn't worked properly.

Apologies again.