/!\ Important Update /!\ r2f
Doing a PROPER revert of my double-take on a past fluke, so differences are clear between r2 and r2f (f for fail).

TLDR: I messed up with versioning between the script I run and the one I publish, and a bad logic mishap made its way on the github version, and I peacefully wrote the comment according to that bad logic, not realizing the code wasn't working properly.

This fix is thus to stop the Ambient Temp dectection from defaulting to step1 or max, with no in-betweens.

I also slipped in just a few tweaks to simplify initial config, since you usually want your max temps to be the same as your last step's.

Sorry for the mess.